src/requirements.txt
# do it in docker, after install new libray
#pip freeze >> requirements.txt
# dev
pytest==6.2.3
requests==2.25.1
#production
alembic
alembic==1.7.5
asyncpg==0.22.0
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
cryptography==36.0.0
databases==0.4.3
#databases[postgresql]==0.4.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.13.6
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.13.4
alembic==1.7.5
asyncpg==0.22.0
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
cryptography==36.0.0
databases==0.4.3
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.63.0
h11==0.12.0
idna==2.10
iniconfig==1.1.1
Mako==1.1.6
MarkupSafe==2.0.1
packaging==21.3
passlib==1.7.4
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
pyparsing==3.0.6
pytest==6.2.3
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
requests==2.25.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.13.6
toml==0.10.2
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.13.4
